# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistama de Logueo",
        "description": "Sistema de prática con los estudiantes de quinto nivel",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/usuario": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "operationId": "4787af24f5968c98d39ff43370d474b5",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombres": {
                                                        "type": "string",
                                                        "example": "Aderson Felix"
                                                    },
                                                    "apellidos": {
                                                        "type": "string",
                                                        "example": "Jara Lazaro"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Registo usuario",
                "operationId": "idcreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "nombre del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Correo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contrasena",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipos_id",
                        "in": "query",
                        "description": "Tipo de usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "402": {
                        "description": "Campos requeridos"
                    }
                }
            }
        },
        "/api/logeo": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "58803ba004ba68fe11c7bde0e77c3d2b",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/usuario/search/{name}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "buscar usuario",
                "operationId": "56fc74dc69bf51f6d234db669cc8eebd",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "ingreso nombre a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "search successful"
                    },
                    "401": {
                        "description": "Invalid search"
                    }
                }
            }
        },
        "/api/usuario/{id}": {
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "eliminar usuario",
                "operationId": "caa57a09cb5aacdd70f3e20843f6eb26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ingreso de id a elminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "destroy successful"
                    },
                    "401": {
                        "description": "Invalid destroy"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}